Working add to shopping cart flow
